This is a big commit, I'm merging two apps developed in another project:
* authapp - authentication handling, this app will show a reCAPTCHA
if the user types the wrong pass more then FAILURE_LIMIT + 1

* registrationapp - user registration app, without mail confirmation.

These two apps share the use of several modules that are kept under
the 'utils' dir in the django project. This is a clear indication
that these apps should also be merged. That will be the next step.

There are several pieces of code by different authors, these are
mentioned where appropriate.